Add errors.ErrorStack(err) for easy stack trace access.
When `New`/`Wrap`/`WrapPrefix` returned `*Error` it was easy to access a stack trace with `errors.Wrap(err, 0).ErrorStack()` in cases where you knew that an error had occurred. Now that `New`/`Wrap`/`WrapPrefix` returns `error` you would have to write `errors.Wrap(err, 0).(*errors.Error).ErrorStack()` to get the same stack trace.

Using the convenience function this becomes simple again: `errors.ErrorStack(err)`. Note that this will panic if err is nil so should only be used when err is known to be not nil.